+++ node/src/test/kotlin/net/corda/node/services/vault/VaultWithCashTest.kt
@@ -9,6 +9,7 @@ import net.corda.core.node.services.Vault
 import net.corda.core.node.services.VaultQueryService
 import net.corda.core.node.services.VaultService
 import net.corda.core.node.services.queryBy
+import net.corda.core.node.services.vault.QueryCriteria
 import net.corda.core.node.services.vault.QueryCriteria.VaultQueryCriteria
 import net.corda.core.transactions.SignedTransaction
 import net.corda.core.transactions.TransactionBuilder
@@ -128,6 +129,7 @@ class VaultWithCashTest : TestDependencyInjectionBase() {
     @Test
     fun `issue and attempt double spend`() {
         val freshKey = services.keyManagementService.freshKey()
+        val criteriaLocked = VaultQueryCriteria(softLockingCondition = QueryCriteria.SoftLockingCondition(QueryCriteria.SoftLockingType.LOCKED_ONLY))
 
         database.transaction {
             // A tx that sends us money.
@@ -138,11 +140,12 @@ class VaultWithCashTest : TestDependencyInjectionBase() {
             println("Cash balance: ${services.getCashBalance(USD)}")
 
             assertThat(vaultQuery.queryBy<Cash.State>().states).hasSize(10)
-            assertThat(vault.softLockedStates<Cash.State>()).hasSize(0)
+            assertThat(vaultQuery.queryBy<Cash.State>(criteriaLocked).states).hasSize(0)
         }
 
         val backgroundExecutor = Executors.newFixedThreadPool(2)
         val countDown = CountDownLatch(2)
+
         // 1st tx that spends our money.
         backgroundExecutor.submit {
             database.transaction {
@@ -154,19 +157,21 @@ class VaultWithCashTest : TestDependencyInjectionBase() {
                     println("txn1: ${txn1.id} spent ${((txn1.tx.outputs[0].data) as Cash.State).amount}")
                     val unconsumedStates1 = vaultQuery.queryBy<Cash.State>()
                     val consumedStates1 = vaultQuery.queryBy<Cash.State>(VaultQueryCriteria(status = Vault.StateStatus.CONSUMED))
+                    val lockedStates1 = vaultQuery.queryBy<Cash.State>(criteriaLocked).states
                     println("""txn1 states:
                                 UNCONSUMED: ${unconsumedStates1.totalStatesAvailable} : $unconsumedStates1,
                                 CONSUMED: ${consumedStates1.totalStatesAvailable} : $consumedStates1,
-                                LOCKED: ${vault.softLockedStates<Cash.State>().count()} : ${vault.softLockedStates<Cash.State>()}
+                                LOCKED: ${lockedStates1.count()} : $lockedStates1
                     """)
                     services.recordTransactions(txn1)
                     println("txn1: Cash balance: ${services.getCashBalance(USD)}")
                     val unconsumedStates2 = vaultQuery.queryBy<Cash.State>()
                     val consumedStates2 = vaultQuery.queryBy<Cash.State>(VaultQueryCriteria(status = Vault.StateStatus.CONSUMED))
+                    val lockedStates2 = vaultQuery.queryBy<Cash.State>(criteriaLocked).states
                     println("""txn1 states:
                                 UNCONSUMED: ${unconsumedStates2.totalStatesAvailable} : $unconsumedStates2,
                                 CONSUMED: ${consumedStates2.totalStatesAvailable} : $consumedStates2,
-                                LOCKED: ${vault.softLockedStates<Cash.State>().count()} : ${vault.softLockedStates<Cash.State>()}
+                                LOCKED: ${lockedStates2.count()} : $lockedStates2
                     """)
                     txn1
                 } catch(e: Exception) {
@@ -188,19 +193,21 @@ class VaultWithCashTest : TestDependencyInjectionBase() {
                     println("txn2: ${txn2.id} spent ${((txn2.tx.outputs[0].data) as Cash.State).amount}")
                     val unconsumedStates1 = vaultQuery.queryBy<Cash.State>()
                     val consumedStates1 = vaultQuery.queryBy<Cash.State>(VaultQueryCriteria(status = Vault.StateStatus.CONSUMED))
+                    val lockedStates1 = vaultQuery.queryBy<Cash.State>(criteriaLocked).states
                     println("""txn2 states:
                                 UNCONSUMED: ${unconsumedStates1.totalStatesAvailable} : $unconsumedStates1,
                                 CONSUMED: ${consumedStates1.totalStatesAvailable} : $consumedStates1,
-                                LOCKED: ${vault.softLockedStates<Cash.State>().count()} : ${vault.softLockedStates<Cash.State>()}
+                                LOCKED: ${lockedStates1.count()} : $lockedStates1
                     """)
                     services.recordTransactions(txn2)
                     println("txn2: Cash balance: ${services.getCashBalance(USD)}")
                     val unconsumedStates2 = vaultQuery.queryBy<Cash.State>()
                     val consumedStates2 = vaultQuery.queryBy<Cash.State>()
+                    val lockedStates2 = vaultQuery.queryBy<Cash.State>(criteriaLocked).states
                     println("""txn2 states:
                                 UNCONSUMED: ${unconsumedStates2.totalStatesAvailable} : $unconsumedStates2,
                                 CONSUMED: ${consumedStates2.totalStatesAvailable} : $consumedStates2,
-                                LOCKED: ${vault.softLockedStates<Cash.State>().count()} : ${vault.softLockedStates<Cash.State>()}
+                                LOCKED: ${lockedStates2.count()} : $lockedStates2
                     """)
                     txn2
                 } catch(e: Exception) {